requirements.txt
flask>=1.0.2
python-dotenv>=0.9.1
sqlalchemy>=1.2.12
flask-sqlalchemy>=2.3.2
psycopg2-binary>=2.7.5
flask-admin>=1.5.2
flask-migrate>=2.2.1
alembic>=1.0.0
gunicorn>=19.9.0
flask-shell-ipython>=0.3.1
bcrypt>=3.1.4
flask-mail>=0.9.1
flask-login>=0.4.1
flask-wtf>=0.14.2
flask-bootstrap>=3.3.7.1
flask-moment>=0.6.0
